TSP using nearest neighbor with a simple happy path test

+++ TravelinSalesmanProblem/TravelinSalesmanProblem/TraveledPoint.h
@@ -7,7 +7,7 @@ public:
     Coordinate loc;
     //Distance to next location
     int dist;
-    TraveledPoint(Coordinate coord, int distance)
+    TraveledPoint(Coordinate coord, double distance)
     {
         loc = coord;
         dist = distance;
